+++ packages/gestalt-codemods/1.22.0-1.23.0/__tests__/toast-remove-color-icon.test.js
@@ -1,0 +1,20 @@
+import { defineTest } from 'jscodeshift/dist/testUtils.js';
+
+jest.mock('../toast-remove-color-icon', () => {
+  return Object.assign(require.requireActual('../toast-remove-color-icon'), {
+    parser: 'flow',
+  });
+});
+
+describe('toast-remove-color-icon', () => {
+  ['toast-remove-color-icon-both', 'toast-remove-color-icon-single'].forEach(
+    test => {
+      defineTest(
+        __dirname,
+        'toast-remove-color-icon',
+        { quote: 'single' },
+        test
+      );
+    }
+  );
+});